fixed ppu sram reading bug

+++ sd5nes/NESHelper.cpp
@@ -7,7 +7,10 @@
 bool NESHelper::GetRandomBool(double trueChance)
 {
 	// Seed the random engine with the current time as seed.
-	static std::mt19937 engine(std::chrono::system_clock::now().time_since_epoch().count());
+	static std::mt19937 engine(static_cast<unsigned int>(
+		std::chrono::system_clock::now().time_since_epoch().count()
+		)
+	);
 
 	std::bernoulli_distribution d(trueChance);
 	return d(engine);
